Added explanation for MAX_LOGLINES
In order to reduce the size of the attachment to below 19.5MB (the maximum for the Red Hat Bugzilla), the maximum lines for any given file is (currently set to) 100,000 lines. If there are any files with more than that, the script will take the last 100,000 lines of that file. This way the attachment is kept to a smaller size while retaining the most recent lines (and hopefully enough relevant information for the bug).

Signed-off-by: Avram Levitter <alevitte@redhat.com>