Limit node-gather pod scheduling time
In order not to exceed overall must-gather time limit, the scheduling of
node-gather pods should be limited to allow collection of data from pods
that were able to get scheduled.

The openshift/origin-must-gather image version is set for 4.2.0, since
in version 4.3.0 the openshift-must-inspect was depracated and causes
failures that should be addressed separatedly.

Signed-off-by: Moti Asayag <masayag@redhat.com>