Improved output for nodes, images and v2v
The must-gather command expects data to be printed out to the pod's log,
and if nothing is received within a given period of time, it will exit
on timeout.  If no data (expecting a newline) is received in the read
call, it may exit with unexpected EOF as well.

To overcome this, this patch speeds up the gathering of node data mainly
by moving all the sriov detection to a single oc exec call, and by
parallelizing gather_nodes.  It also adds some verbosity to
gather_images and removes the expr calls in v2v which displayed lines
containing the result of the expression.

Signed-off-by: Yuval Turgeman <yturgema@redhat.com>