[Release 4.8] Add targeted gathering to collect VM information (#96 and #97)
Don't collect the VM information by default. Instead, add option to collect only VM information for all the VM in the cluster, all the VMs in a specific namespace or for a specific VM by its name.

When running tasks in parallel, it is now possible to override the default number of processes of 5, by setting the `PROS` environment variable.

In addition, added some support for running locally for development, by overriding the default value `BASE_COLLECTION_PATH` environment variable.

Drop the gather_images from the default gathering. It can be still called explicitly.

Split the namespaces out of the main script to avoid using `--all-namespaces` to inspect namespaces.

Remove many duplications.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>