Fix the validate nodes logs functional test
The `Expect(fileInDir(nodeFiles, expectedResource))` code expects nothing. This PR chanes the test to check if there are missing files and list them to the test log, instead of expecting each one of them.

Also, the test now does not check the master nodes as daemon sets are not running there.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>