Fix bug, where the gather_vms_details script fails to collect VM with long names
Fixing BZ-2058925

The bug is where the script is trying to get the vm name from the virt-launcher pod name. The script assumes that the pod name is the vm name with a random addition, separated by a dash. But that is not true for VMs with to long name, and in this case the pod name is not exactly that; e.g. it lack the dash, and the last few characters.

This PR fixes the bug by using different approach to get the VM name from the virt-launcher pod: it reads the `kubevirt.io/domain` annotation in the pod, assuming this is the VM name.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>