e2e tests: improve choosing the right output dir
When running must-gather with --all-images, multiple result directories
are created, one for each must-gather image. The current implementation
of the e2e test takes the first directory it finds. but sometimes it's
the output directory of another must-gather image, and then the test
fails to find the kubevirt must-gather files.

This commit improves the selection of the output directory, by looking
at the first line of the "version" file in the output directory. Since
kv must gather writes a hard coded text to the first line of this file,
we can check if this file was written by kv must gather or not.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>