CNV-70911: Fix the broken CI (#225)
* CI: modify VM to run on creation

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>

* Bump dependencies

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>

* Fix the os-images namespace, and allow its modification

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>

* Fix CNV-71397

Add the `openshift.io/node-selector` missing annotation to
node-gather/node-gather-ds.yaml.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>

* e2e tests: improve choosing the right output dir

When running must-gather with --all-images, multiple result directories
are created, one for each must-gather image. The current implementation
of the e2e test takes the first directory it finds. but sometimes it's
the output directory of another must-gather image, and then the test
fails to find the kubevirt must-gather files.

This commit improves the selection of the output directory, by looking
at the first line of the "version" file in the output directory. Since
kv must gather writes a hard coded text to the first line of this file,
we can check if this file was written by kv must gather or not.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>

---------

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>